+++ tests/integration_tests/test_http_connection.py
@@ -49,12 +49,12 @@ class TestHttpConnection(TestCase):
             self.assertTrue(connection._is_ready())
 
     def test_is_ready_full(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         self.assertTrue(connection._is_ready())
 
     def test_select_one(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"id": "exists", "author": uid, "title": uid, "text": uid})
@@ -64,7 +64,7 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.result[0]["id"], "article:exists", res)
 
     def test_select_all(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid})
@@ -73,21 +73,21 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.status, "OK")
 
     def test_select_one_non_existent(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.select("article", "not_exists")
         self.assertTrue(res.result == [])
         self.assertEqual(res.status, "OK")
 
     def test_select_one_non_existent_table(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.select("not_exists", "not_exists")
         self.assertTrue(res.result == [])
         self.assertEqual(res.status, "OK")
 
     def test_create_one(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         res = connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid})
@@ -100,7 +100,7 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
 
     def test_create_one_with_id(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         res = connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -113,7 +113,7 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
 
     def test_create_one_no_id(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         res = connection.create("article", {"author": uid, "title": uid, "text": uid})
@@ -122,7 +122,7 @@ class TestHttpConnection(TestCase):
         self.assertTrue(res.result["id"] != f"article:⟨{uid}⟩", res)
 
     def test_update_one(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -138,7 +138,7 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.result[0]["author"], "inserted", res)
 
     def test_update_z_all(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -148,7 +148,7 @@ class TestHttpConnection(TestCase):
         self.assertTrue(all(e["author"] == "inserted_all" for e in res.result))
 
     def test_merge_one(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -163,7 +163,7 @@ class TestHttpConnection(TestCase):
         self.assertEqual(res.result[0]["new_field"], "merge", res)
 
     def test_merge_z_all(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -173,21 +173,21 @@ class TestHttpConnection(TestCase):
         self.assertTrue(all(e["all_field"] == "merge" for e in res.result))
 
     def test_query_success(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.query("INFO FOR ROOT;")
         self.assertTrue(len(res.result) > 1)
         self.assertEqual(res.status, "OK")
 
     def test_query_success_with_vars(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.query("INFO FOR ROOT;", variables={})
         self.assertTrue(len(res.result) > 1)
         self.assertEqual(res.status, "OK")
 
     def test_delete_one(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -200,7 +200,7 @@ class TestHttpConnection(TestCase):
         self.assertTrue(len(res.result) == 0)
 
     def test_z_delete_all(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
@@ -212,7 +212,7 @@ class TestHttpConnection(TestCase):
         self.assertTrue(len(res.result) == 0)
 
     def test_export(self):
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.export()
         self.assertTrue("article" in res)
@@ -222,7 +222,7 @@ class TestHttpConnection(TestCase):
 
     def test_import(self):
         file_path = Path(__file__).parent / "import.surql"
-        db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         connection = db.connect()
         res = connection.import_data(file_path)
         self.assertTrue(len(res.result) > 10)
@@ -250,7 +250,7 @@ class TestHttpConnection(TestCase):
 
     # TODO test for let and unset when fix bug https://github.com/surrealdb/surrealdb/issues/3418
     # def test_let(self):
-    #     db = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True, log_level="DEBUG")
+    #     db = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True, log_level="DEBUG")
     #     connection = db.connect()
     #     res = connection.let("name", "LEX")
     #     self.assertEqual(res.result, None)
@@ -265,19 +265,21 @@ class TestHttpConnection(TestCase):
             self.assertFalse(res.is_error())
 
     def test_insert_one(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(27)
             res = connection.insert("article", {"id": uid, "author": uid, "title": uid, "text": uid})
             self.assertFalse(res.is_error(), res)
             self.assertIsNotNone(res.result)
-            self.assertFalse(res.result == [])
+            print(res)
+            # self.assertFalse(res.result == [], res)
             res = connection.select(f"article:{uid}")
             self.assertFalse(res.is_error(), res)
             self.assertIsNotNone(res.result)
+            print(res)
 
     def test_insert_bulk(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(21)
             uid2 = get_random_series(33)
@@ -294,7 +296,7 @@ class TestHttpConnection(TestCase):
             self.assertIsNotNone(res.result)
 
     def test_count(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.count("author")
             self.assertFalse(res.is_error())
@@ -302,21 +304,21 @@ class TestHttpConnection(TestCase):
             self.assertEqual("SELECT count() FROM author GROUP ALL;", res.query)
 
     def test_count_is_zero_if_wrong(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.count("wrong")
             self.assertFalse(res.is_error())
             self.assertEqual(0, res.result)
 
     def test_count_returns_fields(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.count("author:john")
             self.assertFalse(res.is_error())
             self.assertEqual(1, res.result)
 
     def test_db_info(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.db_info()
             self.assertFalse(res.is_error())
@@ -324,7 +326,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual("INFO FOR DB;", res.query)
 
     def test_db_tables(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.db_tables()
             self.assertFalse(res.is_error())
@@ -333,7 +335,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual("INFO FOR DB;", res.query)
 
     def test_session_info(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.session_info()
             self.assertFalse(res.is_error())
@@ -342,7 +344,7 @@ class TestHttpConnection(TestCase):
                               'scope': None, 'session_id': None}, res.result)
 
     def test_remove_table_with_record(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(11)
             tb_name = f"table_{uid}"
@@ -355,7 +357,7 @@ class TestHttpConnection(TestCase):
             self.assertFalse(tb_name in res.result)
 
     def test_remove_table_without_record(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(9)
             tb_name = f"table_{uid}"
@@ -370,7 +372,7 @@ class TestHttpConnection(TestCase):
             self.assertFalse(tb_name in res.result)
 
     def test_insert_when_id_exists_returns_existing(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(11)
             res = connection.insert("article", {'id': uid, 'field': 'old'})
@@ -380,7 +382,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
 
     def test_update_creates_if_not_exists(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(13)
             res = connection.update(f"article:{uid}", {'field': 'old'})
@@ -390,7 +392,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
 
     def test_merge_creates_if_not_exists(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(14)
             res = connection.merge(f"article:{uid}", {'field': 'old'})
@@ -400,7 +402,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
 
     def test_delete_unexisting(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(14)
             res = connection.delete(f"article:{uid}")
@@ -411,21 +413,21 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result, [])
 
     def test_info_root(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.root_info()
             self.assertFalse(res.is_error(), res)
             self.assertEqual(len(res.result), 2)
 
     def test_info_ns(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.ns_info()
             self.assertFalse(res.is_error(), res)
             self.assertEqual(len(res.result), 3)
 
     def test_patch_one(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(24)
             connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid})
@@ -439,7 +441,7 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result[0]['active'], True)
 
     def test_patch_with_diff(self):
-        surreal = Surreal(URL, namespace="test", database="test", credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, namespace="test", database="test", credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             uid = get_random_series(24)
             connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid, "active": False})
@@ -452,13 +454,13 @@ class TestHttpConnection(TestCase):
             self.assertEqual(res.result[0]['active'], True)
 
     def test_is_table_exists(self):
-        surreal = Surreal(URL, 'test', 'test', ('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             self.assertTrue(connection.is_table_exists("person"))
             self.assertFalse(connection.is_table_exists("not_exists"))
 
     def test_info_table(self):
-        surreal = Surreal(URL, 'test', 'test', credentials=('root', 'root'), use_http=True)
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.table_info("author")
             self.assertFalse(res.is_error(), res)